Added PSimpl Douglas-Peuker simplification algorithm Implemented iterator wrapper to present vector<point> as vector<double> in the form x, y, x, y... to use with psimpl algorithm Changed crossThresh type from int to double and defaulted to 0.05 instead of 0

+++ veranda/Packages/pkg_ImageLoader/src/image_loader.cpp
@@ -39,7 +39,7 @@ QVector<WorldObject *> ImageLoader::loadFile(QString filePath, QMap<QString, Wor
     try
     {
         uint64_t colorThreshold = lastOptions->getBlackWhiteThreshold();
-        uint64_t crossThreshold = lastOptions->getCrossProductThreshold();
+        double crossThreshold = lastOptions->getCrossProductThreshold();
         double scaleY = lastOptions->getPxPerHeight();
         double scaleX = lastOptions->getPxPerWidth();
         QColor drawColor = lastOptions->getDrawColor();
